1002,A+B for Polynomials (25)
This time, you are supposed to find A+B where A and B are two polynomials.
Input 
Each input file contains one test case. Each case occupies 2 lines, and each line contains the information of a polynomial:K N1 aN1 N2 aN2 ... NK aNK, where K is the number of nonzero terms in the polynomial,Ni and aNi (i=1,2,...,K) are the exponents and coefficients, repectively. It is given that 1<=k<=10, 0<=NK<...<N2<N1<=1000.
Output
For each test case you should output the sum of A and B in one line, with the same format as the input. Notice that there must be NO extra space at the end of each line. Please be accurate to 1 decimal place.